Add /config volume in docker image
The standard for docker images is to have configuration
related files under /config. This makes it really easy
for end users to setup persistent storage for the config
itself. It also allows for easier use of configmaps in
kubernetes.

This is accomplished by creating a symlink from the
default storage location of BirdNET-GO to the
path /config.