Add /data volume in docker image
The standard for docker images is to have data
related files under /data. This makes it really easy
for end users to setup persistent storage for the data
itself.

This is accomplished by setting the workdir to /data
ensuring that the BirdNET-GO is executed from /data and
storing for example the database there.