Fixes SQL error in GetClipsQualifyingForRemoval
This commit: 16c0961738664f8fc5dbef203d6651258d2c173e did some cleanup
of the GetClipsQualifyingForRemoval and removed a small hack used to
get the main select with the n.<name of column>. Causing the unit
tests to fail. Adds back the AS statement the correct way, utilizing
the built in features of gorm.