feat: add birdnet-user to docker runtime image
By adding the "birdnet-user" for the Docker image, it ensures that the
image is not run as root user. This is considered a best practice
to enhance security and prevent potential vulnerabilities.

Downgrade to less privileged runtime user is done through the entrypoint
script by utilizing gosu: https://github.com/tianon/gosu. Meaning that
the container will be running as user gid=10002, uid=10002 which should
not collide with any existing users on the host system. During the
downgrade process, the volume mount permissions for /data and /config
will be adjusted to the new user.

If the entrypoint user is not privileged, by passing `--user` into the
docker run command, (arbitrary user/group id) the container will be run
as the specified user instead. Here some safe guards are in place to
ensure that provided user/group id has the required permissions to the
mounted volumes as well as belonging to the audio group.