fix: remove redundant copy of reset_auth.sh
For some reason the reset_auth.sh script was copied over to the build
stage, not used there, and then copied over to the runtime stage. This
is redundant and can be removed.

Updates Dockerfile to remove the redundant copy. Instead copies directly
from the host into the runtime stage. In addition, removes the chmod
since the file is already executable.